Update docs on how to identify rotation procedure
All versions of Ops Manager 3.0 include "Rotation Procedure" in the UI
and API for each certificate. Therefore, when directing customers on how
to determine which procedure to use, we can reference this information
instead.

This also removes references to an "Excluded Certificates" section on
the Certificate UI page. This section was an early 2.10 feature that was
removed in 3.0.